Update SetupCrossTenantRelationshipForTargetTenant.ps1
I didn't like the option of creating a text file on GitHub that was just another thing that needs to be updated and could get missed, meaning someone could still be using an older version of the script.

Let me know if this way does not work for everyone.

Also, I fixed the error where you are running the script in a new tenant and have not run Enable-OrganizationCustomization since I hit this issue while I was preparing for my own cross tenant migration and was testing in prod (which is a pain when you have MFA enabled).